+++ renderer/src/data/changelog.js
@@ -1,11 +1,20 @@
+// fixed, improved, added, progress
 export default {
-    description: "Improvements coming soon!",
+    description: "Discord is _still_ making a lot of internal changes!",
     changes: [
+        {
+            title: "Changes",
+            type: "improved",
+            items: [
+                "Plugin startup errors should be more descriptive for developers.",
+            ]
+        },
         {
             title: "Fixes",
             type: "fixed",
             items: [
-                "Injection on Canary is fixed!"
+                "Fixed an issue where custom css crashed Discord.",
+                "Fixed an issue where `waitForModule` returned a boolean instead of a module.",
             ]
         }
     ]