+++ renderer/src/data/settings.js
@@ -20,7 +20,6 @@ export default [
             {type: "switch", id: "store", value: true},
             {type: "switch", id: "autoEnable", value: false},
             {type: "switch", id: "addonErrors", value: true},
-            {type: "switch", id: "autoReload", value: true},
             {type: "dropdown", id: "editAction", value: "detached", options: [{value: "detached"}, {value: "system"}]}
         ]
     },