+++ CHANGELOG.md
@@ -2,6 +2,60 @@
 
 This changelog starts with the restructured 1.0.0 release that happened after context isolation changes. The changelogs here should more-or-less mirror the ones that get shown in the client but probably with less formatting and pizzazz.
 
+## 1.6.3
+
+### Added
+
+### Removed
+
+### Changed
+- Plugin startup errors should be more descriptive for developers.
+
+### Fixed
+- Fixed an issue where custom css crashed Discord.
+- Fixed an issue where `waitForModule` returned a boolean instead of a module.
+
+## 1.6.2
+
+### Added
+
+### Removed
+
+### Changed
+
+### Fixed
+- Fixed non-loading issue due to changed UserSettingsStore
+
+## 1.6.1
+
+### Added
+
+### Removed
+
+### Changed
+
+### Fixed
+- Fixed an issue where `waitForModule` would not return the found module.
+- Fixed an issue where broken addon METAs could prevent BD from fully loading.
+- Fixed an issue where developer badges stopped rendering.
+
+## 1.6.0
+
+### Added
+- Better handling and fallback when the editor fails to load. (Thanks Qb)
+- Now able to sort addons by whether they're enabled. (Thanks TheGreenPig)
+- New `Webpack` API added for plugin developers to take advantage of.
+
+### Removed
+
+### Changed
+- Addon loading no longer uses `require`
+- Addon error modal updated (Thanks Qb)
+- Fixed plugin error display on the modal
+
+### Fixed
+- Fixed dispatcher changes by Discord
+
 ## 1.5.3
 
 ### Added